Reorganize test case by different syntax feature individually
